pyproject.toml
[tool.poetry]
name = "stellar-burger-project"
version = "0.1.0"
description = ""
authors = ["yurapwnz <yurapwnz1992@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.5"
faker = "^37.0.0"
allure-pytest= "^2.13.5"
curlify = "^2.2.1"
playwright = "^1.51.0"
pytest-playwright = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
